Refactor/#91review no other review deleted

+++ review/src/main/java/com/ticketing/review/application/service/ReviewService.java
@@ -146,6 +146,11 @@ public class ReviewService {
         () -> new ReviewException(ErrorCode.REVIEW_NOT_FOUND)
     );
 
+    if (SecurityUtils.getUserRole().equals("ROLE_USER")
+        && SecurityUtils.getUserId() != findReview.getUserId()) {
+      throw new ReviewException(ErrorCode.REVIEW_FORBIDDEN);
+    }
+
     findReview.deleteReview(SecurityUtils.getUserId());
 
     eventPublisher.publishEvent(
